backend-python/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
pandas==2.2.2
black==24.4.2
isort==5.13.2
